+++ src/plot/Plotter.hpp
@@ -26,7 +26,7 @@ namespace {
 class Plotter {
 public:
     // Construct with a hexadecimal plot ID, k parameter, and sub-k parameter
-    Plotter(const std::span<uint8_t const, 32> plot_id, int k, int strength)
+    Plotter(const std::span<uint8_t const, 32> plot_id, uint8_t k, uint8_t strength)
       : plot_id_(to_array(plot_id)), k_(k),
         proof_params_(plot_id_.data(), k_, strength), fragment_codec_(proof_params_), validator_(proof_params_) {}
 
@@ -121,7 +121,7 @@ public:
         // 6) Partitioned Table4 + Table5
         std::vector<std::vector<T4BackPointers>> all_t4;
         std::vector<std::vector<T5Pairing>> all_t5;
-        ProofParams sub_params(plot_id_.data(), proof_params_.get_sub_k(), 2);
+        ProofParams sub_params(plot_id_.data(), numeric_cast<uint8_t>(proof_params_.get_sub_k()), 2);
 
         for (size_t pid = 0; pid < t3_results.partitioned_pairs.size(); ++pid) {
             const auto& partition = t3_results.partitioned_pairs[pid];
@@ -206,7 +206,7 @@ public:
 private:
     // Plot identifiers and parameters
     std::array<uint8_t, 32> plot_id_;
-    int k_;
+    uint8_t k_;
 
     // Core PoSpace objects
     ProofParams proof_params_;
